requirements.txt
pywebview
dnspython
pyinstaller
aiodns